pyproject.toml
[tool.poetry]
name = "pubmed_paper_fetcher"
version = "0.1.0"
description = "Fetch PubMed papers with pharma/biotech authors"
authors = ["Prasritha Narahari <prasrijyo2428@gmail.com>"]
license = "MIT"
packages = [{ include = "pubmed_paper_fetcher" }]

[tool.poetry.scripts]
get-papers-list = "pubmed_paper_fetcher.cli:main"

[tool.poetry.dependencies]
biopython = "^1.83"
python = ">=3.9,<4.0"
requests = "^2.28"
typing-extensions = "^4.0"
pandas = "^2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
